Fix possible crash when browsing a Random page.
This can potentially happen when the task that fetches a Random title
takes a long time, and the PageActivity is stopped in the meantime (this
can be reproduced by adding a delay to RandomArticleIdTask, and checking
"Don't keep activities" in Settings).

Since the task to retrieve a Random title is asynchronous, it means that
the PageActivity might become closed before the task is finished. However,
upon finishing the task, we unconditionally call displayNewPage(),
expecting the activity to be active all the time.

In API 17+ there is a function called Activity.isDestroyed() that can be
used for this purpose, but there's no equivalent function for earlier
APIs... Therefore I added our own isDestroyed() method to our base
Activity class, ThemedActionBarActivity.

Bug: T109281
Change-Id: I8dd624788cdfb14124bb0028d45dfb1352dc0482