Fix image URL replacement for saved pages
* Remove srcsets for remote images.
* Map remote URLs to local files only after all sections are loaded.

The image srcsets attribute should only be specified for remote images
downloaded from the network. For saved pages, only one image is kept
locally and that should be the only image considered. For this reason,
srcsets are now removed when remapping the remote image URLs to local
files. Additionally, this transformation is now performed after all
sections are loaded to avoid a race condition prevalent in long
articles.

Bug: T110487
Change-Id: If2b4642d7507a57085ed0a2185df0d04c1d8432e