Localisation updates from https://translatewiki.net.
A bug in HHVM, T109402, causes most of the line endings to disappear in
translations. After exporting from TWN, the following commands were
executed to printy print the XML:

  find -name strings.xml|
  grep -Ev values/strings.xml|
  xargs -rd\\n -I{} sh -c 'xmllint --format "$1" -o "$1"' -- '{}'

TranslationTests.testAllTranslations() failed because a format specifier
was added to one string (T110243). This failure is a false positive as
it is always safe to add a format specifier, but not usually safe to
remove one without updating all translations since the argument would
not be supplied in the Java code. As a workaround, the tests pass after
running:

  find -name strings.xml|
  grep -Ev "values/|-qq/"|
  xargs -rd\\n sed -ri '/<string name="toast_saved_page"[^>]*>[^<]*<\/string>/d'

Change-Id: I0da2b7f9aaf580d2983315361cfac3e132b73425