Remove redundant field from LinkPreviewFunnel
A page title was being passed to the constructor for LinkPreviewFunnel,
which was then never actually used for anything. I assume this was used
for something in the past. Either way, it's not being used at all now.
This patch removes the field and alters all the calls to the constructor
so that it's not passed any more, to optimise performance and improve
code tidiness.

Change-Id: Ie99666ec6a9f4f0f34a1dc56f7a2e9387b72df4f