Localisation updates from https://translatewiki.net.
TranslationTests.testAllTranslations() failed because a format specifier
was added to one string (T110243). This failure is a false positive as
it is always safe to add a format specifier, but not usually safe to
remove one without updating all translations since the argument would
not be supplied in the Java code. As a workaround, the tests pass after
running:

  find -name strings.xml|
  grep -Ev "values/|-qq/"|
  xargs -rd\\n sed -ri '/<string name="snackbar_saved_page_format"[^>]*>[^<]*<\/string>/d'

Change-Id: I38875b1d87b6cdc9a0373cacf8bd99e7678354bf