Remove LDPI assets
* Remove all LDPI assets.
* Remove LDPI support from convert-icons.py.

It's hard to find LDPI devices these days. The actual LDPI usage metrics
for the app aren't in the Play console but the global Dashboard[0] shows
that the density comprises 3.9% of devices across all versions of
Android. This patch removes LDPI targeted asset cuts and saves our users
some bandwidth.

[0] https://developer.android.com/about/dashboards/index.html#Screens

Bug: T108861
Change-Id: I0db83a94f6dfe9baa9696a952168ad91852c9104