Add exceptions to TranslationTests
* Exclude snackbar_saved_page_format app_store_description from
  translation tests.

TranslationTests.testAllTranslations() failed because a format specifier
was added to one string, snackbar_saved_page_format. This failure is a
false positive as it is always safe to add a format specifier, but not
usually safe to remove one without updating all translations since the
argument would not be supplied in the Java code. As a workaround, the
string has been excluded from the tests until T110243 is fixed properly.

Additionally, app_store_description should not be tested and is excluded
until T91971 is fixed properly.

Change-Id: I489f7a2b93485283847a1f3893ab499d70d07f42