Refine handling of transparent image backgrounds in dark mode
The previous patch submitted on this issue (see
https://gerrit.wikimedia.org/r/#/c/235507/) left at least one instance
in which image colors were being improperly stripped; see Manchester
United F.C. > Kit Evolution.

Hence, just checking the image node and all ancestors for a background-color
style element isn't sufficient.  So with this patch, a white background is
added to each img element that:

(1) Is not nested in a table, unless that table is the infobox and the
element has the class name 'image'; or

(2) Does not have the style property 'background-color' or an ancestor with
it.

And with this, you may examine soccer jersey colors in dark mode to your
heart's content.

Bug: T108333
Change-Id: I350a71d51dd5a452f85c28ef7ef45807455119b1