Fix possible crash in install referrer handler.
This fixes a possible crash when receiving the install referrer intent
after the app gets installed, and also modifies the logic slightly, so
that the event gets logged only if there's at least one nonempty parameter
passed to the receiver.

Based on the data collected so far, it looks like the Play Store often
issues the referrer intent with an empty referrer (i.e. even if the user
hadn't reached the Play Store through a referrer link), so we don't need
to log those events.

Bug: T112101
Change-Id: I1e793b00126a4e6b9c4989fa130fcf0f17c74fbf