Don't hide tables on the main page
As a result of previous refactoring, the hideTables transform was
mistakenly added outside the check to ensure the current page was not
a project main page.  In fact, the transform should only happen off the
main page since main pages often include special, project specific
content and formatting.

This rectifies that.

Change-Id: Ie0ec21dafd3947b55d6b9199981c164a5819e75a