Hide floating action button whilst using find in page
Currently the floating action button remains visible when using find in page.
If the user has activated find in page they've displayed a clear intent such
that the floating action button is redundant. This patch hide the floating
action button whilst the user is using find in page, and makes it visible
again afterwards.

Bug: T111859
Change-Id: Iebfb82e8fd48d33c7707d603204a62b7bb7f0858