Fix sharing of links in Gingerbread.
Our method of removing "blacklisted" intents from the system-provided
intent chooser doesn't seem to work in GB. It's supposed to work by
initializing the chooser with one of the actual intents that will handle
the link sharing, and then adding the rest of the intents using the
EXTRA_INITIAL_INTENTS parameter. Neither of these things seems to have any
effect in GB.

Therefore, in GB, we'll just show the chooser with the blacklisted intents
included (a tradeoff we can live with).

Bug: T111583
Change-Id: Idaedfcbe1f9e576f8a79c3a44bfe78a6034dc8fb