Fix possible crash when dismissing link preview when activity closed.
We call dismiss() as a possible result of an AsyncTask, so if the task
takes a long time, then dismiss() might get called after
onSaveInstanceState() has already been called, which would produce an
exception.  The AsyncTask tries to handle this by checking for isAdded(),
but this seems to keep returning true even after onSaveInstanceState().

This patch overrides the dismiss() function and adds a check for
isResumed(), which actually checks whether this fragment is within the
"live" state (between onResume and onPause).

It may be worthwhile to see if isResumed() should be used instead of
isAdded() in other places, as well.

Bug: T112518
Change-Id: I27cf59b5ccb017600b4b4e9289d94f8c5862aeee