Make references expand/collapse again
An omitted export in collapseTables.js made the function that handles
table collapsing/expanding unavailable to hideRefs.js.  This fixes that.

Also, that function is renamed to describe itself in more accurate, non-
Java-like terms.

There's a further DRY opportunity here in collapseTables/hideRefs but
I'll save that for a follow-up.

Bug: T112208
Change-Id: I7310075defb6bf50c8e81c6ac624efe086665dae