Fix possible crash due to race condition in page loading.
Since we exchange data with the WebView asynchronously, our page loading
logic uses a "sequence" number that our callback functions use to
determine if something is out of sync (e.g. when a new page is loaded
while the previous page hasn't yet finished loading), and this all works
correctly.

...except we also have an asynchronous method of laying out our lead image
container, including the page title and description. The callback for this
functionality did NOT take the sequence number into account, which allowed
for a small window of time where a new page could be loaded while the
layout was in progress, which would throw the loading process out of
whack.

This patch intertwines the sequence number into the lead image layout
callback mechanism, and hopefully closes this remaining gap.

Bug: T112519
Change-Id: I5fcbe1316a5fecad90e1c97ee582df4c082dc063