Add support for loading link preview content via RESTBase
Obtains link preview content from a RESTBase lead section request, if the
article is on English Wikipedia and RESTBase content loading is enabled.

A forthcoming patch will add a feature flag to push this to a small
portion of our beta users.

Also eliminates some unused variables and adds a null check to both
versions of *PageLead.getDescription().

Bug: T111126
Change-Id: I7bb5335639f720fe2fb62422ff2a3cafa93fadfc