Bump minSdkVersion, versionCode, and versionName.
Any patches merged beyond this point shall no longer work in API <15.
This opens the floodgates to the arduous (but extremely pleasant) task
of exfoliating the app of all API-10-specific code.

The minor version number has been bumped by 1, to celebrate the
significance of this occasion.

The versionCode has been bumped by 20, to provide a sufficient gap for
future backporting of features to the Gingerbread version of the app, the
tag for which is located at b97d0cd00f4c4fb932e9fc087334dfe779c59be3.

Change-Id: I06127568bf939585e6f758f5616f418a96d66ff4