No longer strip parenthetical content from first sentence.
This was an experiment to try to improve the readability of the first
sentence of the article, which often contains information in parentheses.
But, since the (relatively naive) algorithm for removing parentheses also
affects links that have parentheses in them, this was doing more harm than
good by breaking these links.

Note: we are still stripping parenthetical content from link previews, but
that's because we don't enable links within the preview. If we end up
enabling links in the preview, then we'll need to revisit this issue.

Bug: T96871
Change-Id: Ic01e82cdf088f6acb42b80f2c355d300907e3950