Prevent TextExtracts from possibly returning wikitext (workaround).
The TextExtracts extension returns content from the lead section of the
requested article. It allows the caller to specify a certain number of
sentences, or a certain number of characters, to retrieve from the lead
section.  But what if the caller requests more data than is present in the
lead section?

Well, in the case of "sentences", the extension correctly returns the
maximum number of available sentences in the lead section, and no more.
However, in the case of "characters", it actually goes beyond the lead
section, and appends content from subsequent sections, which might include
unparsed wikitext (e.g. section headings, edit link, etc).

Previously, we were requesting 512 characters, so that we can parse them
ourselves into two intelligible sentences based on the user's Locale.
This patch changes it to use "sentences": We ask for 10 sentences, and
then re-parse the text using our sentence parsing logic to end up with 2
sentences.

Why do we trust our sentence parsing logic more than TextExtracts? Because
it's better-tailored to the user's Locale on the client side. For example,
in the case of English, the TextExtracts extension incorrectly treats
abbreviations like "i.e.", "B.C.", "Jr.", etc. as separate sentences,
whereas our parser will leave those alone.

Bug: T101153
Change-Id: Id227e82e058debf3efb1fb9aea8f8187b52b2645