Enable a few more syntax highlighting rules when editing.
Syntax highlighting when editing wikitext in the app has been available
for over a year, but has so far been disabled, in the hopes of coming up
with an entirely transformative editing experience in the app. Since we
haven't created such an experience just yet, presenting the user with
modestly highlighted wikitext is better than a solid wall of unhighlighted
wikitext.

Change-Id: I03065408d8b01d82ff262a65c0b957637fbeaafc