Create unique Drawables for mutants
Drawables may be shared unless mutate() is called. When changing
Drawable state, it is safer to create a unique instance by calling
mutate().

This change fixes a bug seen on the AOSP API 23 Nexus 5 and API 15 Nexus
One emulators. I was able to capture repro steps for 23 but not 15:

1 Go the United States article.
2 Set the dark theme.
3 Open a link preview.

Note that the link preview is translucent like the status bar.

SearchBarHideHandler.java
* Call mutate() on stateful Drawables.
* Add missing nullness annotations and final keywords.
* Hold Drawable references instead of Views.
* Remove explicit default values for member variables.
* Add a couple helper methods for getting Resources and colors.
* Replace member variable state, toolbarColor, with stateless method.

GradientUtil.java
* Extract function, setCubicGradient(), from getCubicGradient().

colors.xml
* Update comments.

Change-Id: I6786803bc34f50a88e5a6c267fbd24de238de86f