Beta: turn on/off RB service remotely plus fallback
You can still manually turn on/off RESTBase usage via developer settings.
The roll-out is initially only for the beta app.
The remote config has a percentage of beta app installs which
get to use our RESTBase endpoints for loading page content automatically.

If there is a significant failure it will fall back to using MW API
for the next 100 page loads.
After that it tries again to use RB until there is another RB failure and so on.

Change-Id: Ic189223ebb47f944e8bb80a6b1ec4d6aa98eaed5