Display notification on image save
This patch adds a system notification displayed to the user when an
image is saved from the gallery.  When the main body of the notification
is clicked, it opens the image in the default image viewing application.
When the 'share' button attached to the notification is clicked, a chooser
is displayed for the user to choose an application for sharing.

The Intent to share the image is sent with the image's filename and a link
to a file page with extended information on the image, including license
information.

Bug: T104529
Change-Id: Iea2f133f21190cc2142186df5ad7722383ad473a