Make "use system language" implied
In the interest of avoiding extra finger work and more sophisticated UI,
remove the system language switch and implicitly enable or disable "use
system language" according to the app language entry selected and
the actual system language.

This change might cause confusion for a very small percentage of use
cases. For example, if the user picks Japanese and that language happens
to be the system language, the app language will be implicitly set to
follow the system language, indicated by the special language key value
of null. If the user then leaves the app and changes the system language
to Russian and reenters the app, they're either pleased to see that the
app language automatically updated to Russian without having to wonder
what "system language" meant in the first place OR they're confused and
reset the app language to Japanese again. Thereafter, they never have to
reset the app language even if they change the system language. The
first step is actually unnecessary to arrive at this dilemma since the
initial language is the system language. Despite this potential
confusion, we believe this change will provide the best experience for
most users.

Additional related changes:
* Always return the English language canonical name in
  AppLanguageState.getAppLanguageCanonicalName(). It was a bug to return
  a localized form.
* Add StringUtil.equals() for safe comparison of null Strings.
* Add some missing @NonNull / @Nullable annotations.
* Remove unused string resources.
* Remove some unused methods.
* Update some comments.

Bug: T113097
Bug: T112978
Change-Id: Ia86010b30a7a0976ba9f4a8349616ba495bbf287