Resolve protocol-relative license URLs before passing for external handling
A file's license URL field will sometimes contain a protocol-relative
URL pointing to an on-wiki File: page describing the terms of our use,
particularly in the case of copyrighted media presented under terms of
fair use.

This adds a call to NetworkUtils.resolveProtocolRelativeUrl before
passing them to the system so that it knows how to handle them.

Also refactored the external link handler in GalleryActivity to use
NetworkUtils.resolveProtocolRelativeUrl rather than repeating ourselves.

Bug: T110946
Change-Id: Icfe96b39ddd7e80a37d31626e03949e064fe1f0c