Move the app launcher icons to the mipmap-{density} folder structure
The new convention to store the launcher icon of an android app is to use
the res/mipmap-{...} folders (for example, res/mipmap-xhdpi/launcher.png,
res/mipmap-xxhdpi/launcher.png, and so on). The images in those directories
won't be erased by the most common APK optimization tools (those tools usually
generate a different APK for each display density, and so they remove all the
images under res/drawable-{...} with a density different than the specified).
Since launcher icons can be displayed at a different density than the
device density, it's best practice to put them into the mipmap directories.

Bug: T97239
Change-Id: Iac1bff9e808a41a6b82eb6d97968365c2842a3f0