6.0 runtime permission for save image from gallery
Added runtime permission checks before trying to write to
external storage in gallery.

Note this patch does not address the share a fact functionality since
it doesn't currently work, see T116120.

The runtime checks are used once target SDK level is set to 23.
I'll let a later patch change this value, once we're done with all the
permission checks and the backup check.

Bug: T113750
Change-Id: I356a736a0d84c99e05e0a6ae1784f99c1c59779e