Fix various bugs to improve the offline browsing experience
Fix the logic so the app always tries to load the requested page
by all means possible, so if the network connection is not
available it will try to fetch it from the cache or from the saved
pages storage. This patch changes a lot of code from the
JsonPageLoadStrategy, including the "reload" logic.
The Link Previews are also modified so that if the preview
couldn't be retrieved from the network but there is a
cached version of the page in the device, the link preview
will be constructed from the page contents.

Bug: T85065
Change-Id: Ie335dc94b3d0ee3f94d190900d885746aa118465