Fix "add tab" icon for xxxhdpi.
This icon was not generated correctly by the rsvg tool, and looked weird.
We've observed this behavior before, where certain SVG images would not
render correctly by rsvg. Therefore I'm deleting the SVG file for this
icon, and instead using the pre-generated PNG files provided directly from
the Material icon set.

We should probably do this for other Material icons that we use, as well,
although the others look correctly-generated.

Change-Id: I87972f62f943a379a20928405fddf76408d310bb