Add Butter Knife and upgrade Checkstyle
The Android app wants to do more with native Views. In this interest,
Butter Knife v7.0.1 is added, a tool that is helpful in eliminating some
of the boilerplate of native Views.

BK increases the size of the dev release apk from 4201830B to 4220733B,
about ~18KiB.

To verify that the library is working properly, one usage of BK is made
in ArticleHeaderView.

BK uses annotations and requires annotated fields to be non-private
which violates Checkstyle rules. The implicit Checkstyle version was
v5.9[0] which does not support the ignoreAnnotationCanonicalNames
property of the VisibilityModifier module. This patch upgrades
Checkstyle to the latest, v6.11.2, adds the BK @Bind annotated property
exception, and fixes all new violations of type "Redundant 'public'
modifier".

[0] https://github.com/gradle/gradle/blob/master/subprojects/code-quality/src/main/groovy/org/gradle/api/plugins/quality/CheckstylePlugin.groovy#L21

Change-Id: I08af3779dba038f4e349a68f1e600782cda1fc52