6.0 disable backup + bump target SDK to 23
Had issues testing the backup and restore of app data.
So, I'm disabling it for now. (See earlier PatchSet for details.)

Also bumping target SDK level to 23 since this should be the
last step to get ready for Marshmallow. (The runtime permission
code for share-a-fact should come when that gets re-enabled
for 6.0 devices).

Bug: T116232
Change-Id: I12bbf6dea7b366b3da7a6d3707d642e241812169