Fix saved pages with spaces in their title not being recognized as saved
When following a link from a page to another, it's checked if that page
is saved in disk. The problem was, it was stored with the "Display title"
(with spaces) and the table was queried with the "Normal_Title", with
underscores. The fix is to make sure that the PageTitle.getText
method *allways* returns the title with underscores.
Included in this change is also a database migration which
enforces all the titles stored in the DB to use the unserscore
format.

Bug: T108885
Change-Id: I22928abc7dec3bfecceea22c3f6c7387dfbc927e