Add share-a-fact support on 6.0 Marshmallow
-Adds support for injecting our custom CAB menu via onActionModeStarted/
Finished on 6.0+, since onSupportActionModeStarted/Finished are no longer
called.

-This required finding a way to prevent the custom CAB menu from
being injected during a long press in the search bar.  It turns out this
can happen because of new custom insertion action mode functionality
introduced in Marshmallow.  Unfortunately, the insertion/selection logic
is hidden entirely in private fields in the TextView class, and the
ActionMode API does not provide a way of finding the source view or action
triggering the action mode, nor whether the action mode is a selection
action mode or an insertion action mode.  So, (h/t @dbrant) I added a
getOriginatingView utility method to obtain the originating view of the
action mode by reflection.

-Additionally, updated the share-a-fact functionality to go through the
content provider framework on Marshmallow and above to prevent
permissions-related errors when sharing a fact as an image.[0]

[0] See https://stackoverflow.com/questions/32981194/android-6-cannot-share-files-anymore

Bug: T116120
Change-Id: I48f5a99d5c14a4a53d7f12abec01abb291268fee