Restrict WebView debug enabled call to debug APKs only
Only call WebView.setWebContentsDebuggingEnabled
if app is built in debug mode. No need to call this for release
builds since the default is false.

This patch may or may not fix T113052 but it will change the kind
of outer stack traces we get for this issue (the inner most part, though,
should stay the same). It will postpone the missing WebView issue
until later. The call happens very early when the Application object is created.

Bug: T113052
Change-Id: Ia9bd9869f7f5fc8004782d996cb4d379e69d8b24