Fix possible unnecessary tab creation.
When the app is launched via a VIEW intent (e.g. from an external
Browser), and it didn't have any tabs open from previous sessions, the app
unconditionally creates a new tab for the Main Page *and* a new tab for
the page requested in the VIEW intent, resulting in a puzzling appearance
of a new tab on top of a "blank" tab.

This patch makes it so that the app creates a new tab only if the topmost
tab's backstack is nonempty. This successfully handles the case described
above.

Bug: T116502
Change-Id: I6557678d11e965fd28f63c76e835aa85317c5075