Disable WebView hardware acceleration
Enabling hardware acceleration is causing a variety of Exceptions across
API levels, mostly KitKat, v4.4.4, API 19. Since enabling acceleration
was originally intended to fix T109983 but didn't, this removes the
setting, restoring the default behavior of software rendering. An
alternative solution might be to catch any Throwable emitted when
enabling acceleration but this might put the object in a bad state:

  @Override
  public void setLayerType(int layerType, Paint paint) {
    if (ApiUtil.hasLollipop()) {
      try {
        super.setLayerType(layerType, paint);
      } catch (Throwable e) {
        L.w(e);
      }
    }
  }

Change-Id: Ib5f812ed847e7dcf1905a95133db42d4de2a448f