A/B test of two link preview designs.
This implements an a/b test of two designs for the link preview dialog:
A) The current design.
B) The original "prototype 1" design, which includes an excerpt of text
with the lead image side by side, and no overflow menu (and no thumbnail
gallery).

The hope is to gauge user response to a trimmed-down, back-to-basics
design, and see how it affects engagement.

Bug: T117065
Change-Id: Ic65ac96106b3949ce121ba7c86d4dda68bf2ea8f