Disable some DOM transformations for RESTBase users
The DOM transformations that can be handled by the service
are handled by it (I9437791ba8f5953be98546f76a6d15437cb58fc2).
Thus these don't need to be run again on the client.
Esp. moving the first paragraph around twice would be undesirable.

So we can better keep track what has been moved to the service
I moved the file that implement these transformations to a service
sub-folder.

I've added small notes at the end of the call to the respective DOM
transformations which have not been moved to the service (yet),
indicating the category of the reason why that has not yet been done.
Theses reasons are:
* we need to add a clickHandler,
* it's a dynamic client setting (dark mode), or
* the transformation depends on the device width.

 I think the clickHandler transformations could
 potentially be split up into two parts. The pure DOM transformation
 and the addition of the clickHandler. We then could move the first
 part tot he service, too.

Bug: T111443
Change-Id: Ibf76222d0ba10520797397fdbd95205c9f7d995b