Implement event logging for page scrolling.
This implements a new funnel to log how far the user scrolls through any
given article. It logs the total "flux" of scrolling done by the user in
the up or down direction, as well as the total depth that the user reached
into the article.

This also updates the ToC funnel to include additional data for free, such
as the page ID, and the index and title of the selected section.

Bug: T116352
Change-Id: Id8f6246b9c063af3cd7e32e0a0f9d79eb9878577