Navigate to target URL specified in Install Referrer.
This allows the app to automatically open the referring URL specified in
the INSTALL_REFERRER intent. This makes it possible to show a banner in
mobile web that takes the user to the Play Store to install the app, and
then for the app to automatically go to the page that the user was
browsing in mobile web.

Bug: T117219
Change-Id: I5fdf38ee6f0c3ea4cf787a38ec9bfa981fed88cf