Upgrade to Mapbox GL SDK.
The good news: This brings the app up to date with the latest SDK from
Mapbox.

The bad news: This brings our APK size from 4.1MB to 14.4MB (!)
A sign of the times :(

This patch includes somewhat of a redesign, or more accurately a bit of a
stripdown of the interface, compared with the previous design. This will
now be a simple full-screen map with markers. Clicking on a marker brings
up a link preview for the given article. Nice and simple.
This means that the "list" interface with accelerometer-based arrows has
been removed, pending some new designs for a location-aware link preview.

Also updated the marker image that pinpoints to articles on the map.

Bug: T117803
Change-Id: I3dc224b02cbbb894ee263525e081fcfe38f9e1d3