Hygiene: move lead parallax to Drawable matrix
The parallax effect used by the lead image was accomplished by sizing
its housing ImageView to the full dimensions of the image and then
translating the View itself to achieve a parallax effect. This was nice
because it avoided math and translations are cheap but poor because it
changed layout parameters without concern for adjacent Views. This
unusual behavior made it difficult to visualize the View hierarchy.

This patch moves the parallax effect from raw View resizing and
translations to the ImageView's Drawable matrix. The math is a little
intimidating in some parts, but it's mostly pure (free of implementation
details), and the View is now boring and predictable. This change will
make it easy to add or remove Views without surprising things happening.

Bug: T116122
Change-Id: I17ae6c24f876a37b1c86180daf4794dfaa2696bd