Update SDK Manager Plugin to d0c113b snapshot
There hasn't been a release of the SDK Manager Plugin for a while. The
latest release does not properly download ProGuard dependencies[0] which
requires us to use the Jenkins Android Emulator Plugin to download them
for CI. Since the Android Emulator CI server is very slow, a dependency
on it is impractical for jobs that otherwise would not need it. Use
JitPack to grab the SDK Manager Plugin's master HEAD, d0c113b.

[0] https://github.com/JakeWharton/sdk-manager-plugin/issues/57

Change-Id: Ia0f5f8dc64a676bdead0bd79b36b8071d45cb516