Fix clipped text in alpha and beta "About" page
Presently, in the alpha and beta apps, the title of the app is changed in
the "About" page, but since the height of the TextView and LinearLayout are
defined absolutely, the text gets clipped on narrow devices (like most
phones). This patch changes this by letting the TextView and LinearLayout
height wrap the content, but specificies a minimum size the same as the old
actual size in order to preserve layout.

Bug: T119359
Change-Id: I4079a0a842a8b3394a42082aebdf4006503fc77c