Location-aware link previews.
This adds a button to "Get directions" to the link preview that pops up
when tapping on a map marker. The button fires an Intent with a "geo:" Uri
that should be picked up by the default Maps app installed on the device.

Note: this patch does not implement the "Get directions" menu item in the
link preview's overflow menu (to be shown when the link preview is invoked
from a regular link in an article that has a location). This will be done
in a subsequent patch.

Bug: T119215
Change-Id: I9d59321e89982d79c8255960a01325259d1b3d9b