Fix infinite search loop
- Only clear results when the query changes and the new result set is
  received.
- Add @Nullable / @NonNull annotations where obvious.

The search results Fragment has several sources of input including:

1 Prefix search
2 Full text search
3 Search suggestion
4 Cache

There are special cases but the primary flow intents for starting a new
search (whenever query changes) appear to be:

1 Search term is identical: do nothing.
2 Search term is empty: clear results.
3 Cache is nonempty: clear old results and show new.
4 Cache is empty: perform a prefix and suggestion search.
5 Prefix and suggestions search was nonempty: clear old results and show
  new, clear or show the search suggestion too.
6 Prefix and suggestions search was empty: clear or show suggestion and
  perform a full text search.
7 Full text search completed: clear old results and show any new.

Additionally, when the end of the results list is scrolled to, full text
searches are performed and the results are appended. This seems to have
been the source of the issue as results were being cleared instead of
appended, so the user was always at the bottom of the list and always
triggering another search.

Bug: T120148
Change-Id: Iee4107261a572f2e20943eca92f43ff290190520