Temporarily lessen tab unmarshal failure penalty
For nonproduction builds, temporarily allow problematic tabs to be
discarded and tabs in the new 6191e5e format to be written.

Per 59dce5f, in nonproduction builds the penalty for remote logging
should be death. The thinking was that if it's worth burning user data
and engineering parsing time, it's worth crashing on everything but
prod. A TODO has been added to revert this patch when the next beta,
2.1.135 is published.

Change-Id: I86d08d4a48c591b7a2a415ab92b0b6bcfe46f069