Fix article and menu bar padding on tablets.
On a tablet screen, the article top-padding was too much, since it was
using the *horizontal* padding dimen, which grows to 128dp for a tablet
screen. It should always be a fixed 16dp from the top.

Conversely, the padding to the left and right of the new native menu bar
should in fact be equal to the dimen for the horizontal activity margin.

Here's a "before" shot:
https://veuwer.com/i/3g2p

And after:
https://veuwer.com/i/3g2q

Change-Id: I32a8177d5bfdbd39076e0ecf6a85aca5d10e3fa5