Fix share-a-fact cards when lead image scrolled away.
Share-a-fact cards were no longer showing the background image correctly
because the underlying ImageView is now using a Matrix transform to
achieve the parallax effect, and it becomes incorrectly applied to the
share-a-fact card.

This patch simplifies things by letting the share-a-fact card access the
underlying Bitmap of the ImageView, instead of making the ImageView draw
itself onto a new Bitmap.

In the previous implementation, I see there was some concern regarding a
possible exception due to using a recycled bitmap, but I don't see how
this can happen, given our current way of drawing the card.

Bug: T123283
Change-Id: Iff4731499fe324f632c42543eeaa1b1ff157fddb