Kickin' it with Fresco.
NB1: This removes Picasso as well as PhotoView, because PhotoView does
not support Fresco-based views, so it can no longer provide the pinch-to-zoom
functionality that we so crave.
In a subsequent patch we'll restore pinch-to-zoom in the Gallery using a
different means.

NB2: This also restricts the native ABIs that get packaged into the APK,
since Mapbox doesn't come with arm64 libs, which causes link errors. Once
Mapbox is updated for arm64, we should remove the restriction.

Bug: T124733
Change-Id: I1edf132ee8a3ee50321bd5ab4c8ded9f1b1daad1