Retain preload channel ID
We now keep the same package ID for OEM installs as the Play Store. This
is good for users and developers because everyone gets the same apk when
updates are made. This is bad for trackers because the preload channel
ID, a property of the original preloaded apk and different for upgraded
apks, was lost as soon as the phone was connected to the internet and
upgrades are downloaded. We now record the channel ID on first boot.

Bug: T115396
Change-Id: Iec363ee4b011dd70e457dcea733fc21551244def