Prevent crash after requesting empty definition from prod content service
Sending an empty string to the production content service to request
a Wiktionary definition causes the app to crash (though oddly enough this
does not happen when running locally).  This adds a check to immediately
fire a "no definitions found" dialog rather than sending the request when
selectedText.trim() is empty.

Also added a minor fix to hide the progress bar consistently when the "no
definitions found" message is displayed.

Bug: T123783
Change-Id: Iba04e02e16ca5cd1d7d8723999f618c8995fa761