Don't update the article header for the error page
The article header should not be shown for the error page. The header
was mistakenly updated on configuration change which caused the page to
display as an odd hybrid of error page and the previous wiki page
visited. This patch amends that behavior to only update when not on the
error page.

Bug: T120279
Change-Id: I2b35a6880aca9eda6805d0af1824497b855e22ef