Don't use javascript bridge for copying to clipboard.
When passing the text selection from the WebView to the Java layer, we
limit the number of characters on the WebView side, which is not a bad
thing, because a large enough selection can lead to an OOM error.

However, for the purpose of copying the selection to the clipboard, it can
be a hindrance to limit the number of characters, and is confusing to the
user. This patch copies the selected text by simulating a "Ctrl-C"
keyboard event, and dispatches it to the WebView, which causes it to copy
the selection to the clipboard natively.

Additionally, this allows the copied text to preserve any *formatting*
and/or links, instead of simply copying plain text.

Change-Id: I935cf1d9dc8ac78650b76c6e8cfb758114dea4ac