Hygiene: rename PersistenceHelper to DatabaseTable
There's a clear line of distinction between a ContentPersister and a
PersistenceHelper. A ContentPersister uses a ContentProviderClient to
interact with the underlying database table. A PersistenceHelper
interacts directly with the database to, for example, create a specific
table, upgrade a specific table, operate on a specific table's rows and
columns, and defines a Column nested class and the version of the
database that the table was added. Rename PersistenceHelper to
DatabaseTable so it's clear it has a lot to do with a specific table in
the database. No functional changes intended.

Change-Id: If80d7ea792a907b278e0894cf23c9694b605037d