Add stub authenticator
The sync adapter's account is user visible even if the sync adapter
itself is not. This means users can add or remove a Wikimedia account
from the system account settings screen. This patch enables that
functionality but doesn't provide an account that is usable by other
apps, including other app flavors (dev, alpha, beta, prod), which is a
future improvement.

Change-Id: I0b562771d89787b589a1273bed87505c88cbd29c