Log action extlink-always when user switches off Zero exit warnings
This is a tentative implementation of the extlink-always action in
https://meta.wikimedia.org/wiki/Schema:WikipediaZeroUsage.

Apparently, in the exit intertitial dialog on the mobile web, there was
once an option to suppress further warnings and always continue.  I think
this is what Yuri had in mind when creating the schema.  But I can't
replicate that behavior.

If we decide to include such an option in our exit warning dialog, we can
log this action there, too, but manually switching off the warnings in the
settings also seems like an action that meets the criteria for logging
extlink-always.

Bug: T123748
Change-Id: I69ef8cf7dacf0266a4f51d5bf81c706434d6fb40